+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1507,7 +1507,7 @@ class Evaluations3(BaseModel):
         pydantic.Field(
             deprecated="warning: ** DEPRECATED ** - This will be removed in a future release, please migrate away from it as soon as possible."
         ),
-    ] = parse_datetime("2026-06-17T11:59:36.977Z")
+    ] = parse_datetime("2026-06-17T13:32:44.271Z")
     r"""Deprecated. The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1652,7 +1652,7 @@ class Evaluations2(BaseModel):
         pydantic.Field(
             deprecated="warning: ** DEPRECATED ** - This will be removed in a future release, please migrate away from it as soon as possible."
         ),
-    ] = parse_datetime("2026-06-17T11:59:36.976Z")
+    ] = parse_datetime("2026-06-17T13:32:44.270Z")
     r"""Deprecated. The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1797,7 +1797,7 @@ class Evaluations1(BaseModel):
         pydantic.Field(
             deprecated="warning: ** DEPRECATED ** - This will be removed in a future release, please migrate away from it as soon as possible."
         ),
-    ] = parse_datetime("2026-06-17T11:59:36.975Z")
+    ] = parse_datetime("2026-06-17T13:32:44.269Z")
     r"""Deprecated. The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1902,7 +1902,7 @@ class CreateDatasetItemResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-17T11:59:16.603Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-17T13:32:19.477Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")