+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVAQ6ZXGY0CQHREYD6E5E7E9"
+        "01KVAWHG90QWP7Y6RWZ2NBTCHG"
     )
     r"""The unique identifier of the data source"""