+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVAQ6ZXH3234HF2C692C58RG"
+        "01KVAWHG9273ZN77XH64Z0WR5W"
     )
     r"""The unique identifier of the data source"""