+++ packages/orq-rc/RELEASES.md
@@ -8799,3 +8799,13 @@ Based on:
 - [python v4.11.0-rc.29] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.29] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.29 - packages/orq-rc
+
+## 2026-06-17 13:35:15
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.781.0 (2.907.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.30] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.30] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.30 - packages/orq-rc